extract router
