+++ Makefile
@@ -8,4 +8,4 @@ test-cover:
 	go test -coverprofile=coverage.out ./...
 	go tool cover -html=coverage.out
 generate-doc: ## Generate Swagger Api Documentation
-	swag init --parseDependency --parseInternal --parseDepth 1
+	cd cmd/api/ && swag init --parseDependency --parseInternal --parseDepth 1 -o ../../docs
