+++ internal/app.go
@@ -1,18 +1,14 @@
 package internal
 
 import (
-	"github.com/gin-gonic/gin"
 	"github.com/joho/godotenv"
-	docs "github.com/mateus-sousa-dev/meet-people/docs"
 	"github.com/mateus-sousa-dev/meet-people/internal/emails"
 	"github.com/mateus-sousa-dev/meet-people/internal/events"
 	"github.com/mateus-sousa-dev/meet-people/internal/infra"
 	"github.com/mateus-sousa-dev/meet-people/internal/login"
-	"github.com/mateus-sousa-dev/meet-people/internal/middlewares"
 	"github.com/mateus-sousa-dev/meet-people/internal/passwordresetconfigs"
+	"github.com/mateus-sousa-dev/meet-people/internal/routes"
 	"github.com/mateus-sousa-dev/meet-people/internal/users"
-	swaggerfiles "github.com/swaggo/files"
-	ginSwagger "github.com/swaggo/gin-swagger"
 	"log"
 )
 
@@ -49,20 +45,5 @@ func StartApplication() {
 	userDelivery := users.NewDelivery(writingUseCase)
 	loginUseCase := login.NewLoginUseCase(userRepository)
 	loginDelivery := login.NewDelivery(loginUseCase)
-	SetupRoutes(userDelivery, loginDelivery)
-}
-
-func SetupRoutes(userDelivery users.Delivery, loginDelivery login.Delivery) {
-	r := gin.Default()
-	docs.SwaggerInfo.BasePath = "/"
-	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerfiles.Handler))
-	apiV1Routes := r.Group("/api/v1")
-	apiV1Routes.POST("/users", userDelivery.CreateUser)
-	apiV1Routes.GET("/activate-account/:activationpath", userDelivery.ActivateAccount)
-	apiV1Routes.PATCH("/forgot-password", userDelivery.ForgotPassword)
-	apiV1Routes.GET("/validate-url-password/:urlpasswordreset", userDelivery.ValidateUrlPassword)
-	apiV1Routes.PATCH("/reset-forgotten-password/:urlpasswordreset", userDelivery.ResetForgottenPassword)
-	apiV1Routes.POST("/login", loginDelivery.Exec)
-	apiV1Routes.GET("/logged", middlewares.Authenticate(userDelivery.Logged))
-	r.Run()
+	routes.SetupRoutes(userDelivery, loginDelivery)
 }
